Move drawing in separate class
